# internal/docs/swagger.yaml
definitions:
  dto.SubscriptionDTOSwagger:
    properties:
      end_date:
        type: string
      price_per_day:
        type: number
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  dto.SubscriptionJSON:
    properties:
      end_date:
        description: MM-YYYY
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        description: MM-YYYY
        type: string
      user_id:
        type: string
    type: object
  swagger.HTTPErrorDTO:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /subscriptions:
    get:
      description: Возвращает все подписки пользователя, можно фильтровать по user_id
        и service_name
      parameters:
      - description: User UUID
        in: query
        name: user_id
        type: string
      - description: Service name
        in: query
        name: service
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SubscriptionDTOSwagger'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.HTTPErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.HTTPErrorDTO'
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создает новую подписку
      parameters:
      - description: Subscription payload
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/dto.SubscriptionJSON'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SubscriptionDTOSwagger'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.HTTPErrorDTO'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/swagger.HTTPErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.HTTPErrorDTO'
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Удаляет подписку по ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.HTTPErrorDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swagger.HTTPErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.HTTPErrorDTO'
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      description: Возвращает подписку по ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionDTOSwagger'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swagger.HTTPErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.HTTPErrorDTO'
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет подписку по ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Subscription payload
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/dto.SubscriptionJSON'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionDTOSwagger'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.HTTPErrorDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swagger.HTTPErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.HTTPErrorDTO'
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/costs:
    get:
      description: Возвращает суммарную стоимость подписок за период с опциональными
        фильтрами
      parameters:
      - description: Start date (DD-MM-YYYY)
        in: query
        name: from
        required: true
        type: string
      - description: End date (DD-MM-YYYY)
        in: query
        name: to
        required: true
        type: string
      - description: User UUID
        in: query
        name: user_id
        type: string
      - description: Service name
        in: query
        name: service
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.HTTPErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.HTTPErrorDTO'
      summary: Get total subscriptions cost for period
      tags:
      - subscriptions
swagger: "2.0"